Move mapping_quality to maponly
