+++ src/mapper.rs
@@ -854,20 +854,6 @@ fn is_proper_pair(a1: &Alignment, a2: &Alignment, mu: f32, sigma: f32) -> bool {
     same_reference && insert_good && rel_orientation_good
 }
 
-/// Return mapping quality for the top NAM
-pub fn mapping_quality(nams: &[Nam]) -> u8 {
-    if nams.len() <= 1 {
-        return 60;
-    }
-    let s1 = nams[0].score;
-    let s2 = nams[1].score;
-    // from minimap2: MAPQ = 40(1−s2/s1) ·min{1,|M|/10} · log s1
-    let min_matches = min(nams[0].anchors.len(), 10) as f32 / 10.0;
-    let uncapped_mapq = 40.0 * (1.0 - s2 / s1) * min_matches * s1.ln();
-
-    uncapped_mapq.min(60.0) as u8
-}
-
 /// A scored alignment pair
 #[derive(Debug, Clone)]
 pub struct ScoredAlignmentPair {
